+++ src/models/entity_pedido.py
@@ -1,6 +1,6 @@
 # Tabela Pedido:
 
-from sqlalchemy import Column, Integer, Float, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, Float, DateTime, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 from datetime import datetime, timezone
 from src.database.base import Base
@@ -9,7 +9,7 @@ class Pedido(Base):
     __tablename__ = 'pedidos'
     id_pedido = Column(Integer, primary_key=True)
     data_pedido = Column(DateTime, default=datetime.utcnow)
-    valor_total = Column(Float, nullable=False)
+    valor_total = Column(Numeric(10,2), nullable=False)
     # Chave estrangeira da entidade Cliente por ter relação direta:
     id_cliente = Column(Integer, ForeignKey("clientes.id_cliente"),nullable=False)
     # Relacionamentos:
